Added block times Line each 30 min default
