+++ src/DayView.js
@@ -1,9 +1,10 @@
 // @flow
-import { View, Text, ScrollView, TouchableOpacity, RefreshControl } from 'react-native';
-import React, { useEffect, useState } from 'react';
+import { View, Text, ScrollView, TouchableOpacity, Platform } from 'react-native';
+import React, { useEffect, useRef } from 'react';
 import moment from 'moment';
 import _ from 'lodash';
 import populateEvents from './Packer';
+import {MINUTES_PER_BLOCK, BLOCK_HEIGHT, HEIGHT_PER_MINUTE} from './constants';
 
 const LEFT_MARGIN = 60 - 1;
 // const RIGHT_MARGIN = 10
@@ -30,64 +31,69 @@ const DayView = ({
   styles,
   width,
 }) => {
-  const calendarHeight = (end - start) * 100;
   const containerWidth = width - LEFT_MARGIN;
-  const packedEvents = populateEvents(events, width, start);
+  const blockedEvents = events.filter((e) => e.booking_type === 'blocked');
+  const normalEvents = events.filter((e) => e.booking_type !== 'blocked');
+  const packedEvents = populateEvents(normalEvents, containerWidth, start);
+  const scrollViewRef = useRef(null);
 
-  const [refreshing, setRefreshing] = React.useState(false);
+  const contentOffset = () => {
+    const nowHour = moment().hour();
+    if ((nowHour < (start + 1)) || (nowHour > (end + 1))) return 0;
+    return nowTop() - 100;
+  };
 
-  const onRefresh = React.useCallback(() => {
-    setRefreshing(true);
+  useEffect(() => {
+    if (Platform.OS === 'ios') return;
+    setTimeout(() => {
+      if (!scrollViewRef.current) return;
+      scrollViewRef.current.scrollTo({ y: contentOffset(), animated: false });
+    }, 0);
+  }, []);
 
-    wait(2000).then(() => setRefreshing(false));
-  }, [refreshing]);
+  const nowTop = () => {
+    const now = moment();
+    const nowInMinutes = now.hour() * 60 + now.minutes();
+    const startInMinutes = start * 60;
+    return ((nowInMinutes - startInMinutes) / MINUTES_PER_BLOCK) * BLOCK_HEIGHT;
+  };
 
   const renderRedLine = () => {
     const now = moment();
     if (!now.isSame(date, 'day')) return null;
-    const offset = 100;
-    const timeNowHour = now.hour();
-    const timeNowMin = now.minutes();
+    const top = nowTop();
+    const lineWidth = width - 20;
     return (
-      <View
-        key="timeNow"
-        style={[
-          styles.lineNow,
-          {
-            top:
-              offset * (timeNowHour - start)
-              + (offset * timeNowMin) / 60,
-            width: width - 20,
-          },
-        ]}
-      />
+      <>
+        <View style={[styles.lineNow, { top, width: lineWidth }]} />
+        <View style={[styles.circleNow, { top: top - 3 }]} />
+      </>
     );
   };
 
   const renderLines = () => {
-    const offset = calendarHeight / (end - start);
-    const time = moment().startOf('date');
+    const time = moment().hour(start).minute(0);
+    const lines = (((end - start) * 60) / MINUTES_PER_BLOCK);
 
-
-    return [...Array(24).keys()].map((i) => {
+    return [...Array(lines).keys()].map((i) => {
       const timeText = format24h ? time.format('HH:mm') : time.format('HH:mm A');
-      time.add(1, 'hour').format();
+      time.add(MINUTES_PER_BLOCK, 'minutes');
       return (
-        <>
-          <Text style={[styles.timeLabel, { top: offset * i - 6 }]}>
+        <View key={timeText}>
+          <Text style={[styles.timeLabel, { top: BLOCK_HEIGHT * i - 6 }]}>
             {timeText}
           </Text>
-          <View style={[styles.line, { top: offset * i, width: width - 20 }]} />
-        </>
+          <View style={[styles.line, { top: BLOCK_HEIGHT * i, width: width - 20 }]} />
+        </View>
       );
     });
-
   };
 
   const eventTappedHandler = (event) => { onEventTapped(event); };
 
   const renderEvents = () => {
     const componentEvents = packedEvents.map((event, i) => {
+      if (event.booking_type === "blocked") return null;
       const style = {
         left: event.left,
         height: event.height,
@@ -97,6 +103,7 @@ const DayView = ({
 
       const eventColor = {
         backgroundColor: event.color,
+        boderColor: event.boderColor,
       };
 
       // Fixing the number of lines for the event title makes this calculation easier.
@@ -142,11 +149,69 @@ const DayView = ({
     );
   };
 
+  const rotatedLenght = (lenght) => (Math.sqrt(2 * (lenght ** 2)));
+  
+
+  const renderBlocks = () => {
+    const lineWidth = 4;
+    const lineMargin = 8;
+    return blockedEvents.map((event) => {
+      const startTime = moment(event.start);
+      const endTime = moment(event.end);
+      const dayStartTime = startTime.clone().hour(start).minute(0);
+
+      const top = startTime.diff(dayStartTime, 'minutes', true) * HEIGHT_PER_MINUTE;
+      const height = endTime.diff(startTime, 'minutes', true) * HEIGHT_PER_MINUTE;
+
+
+      const rotatedLineWidth = rotatedLenght(lineWidth);
+      const rotatedLineMargin = rotatedLenght(lineMargin);
+      const rotatedHeight = rotatedLenght(height);
+
+      const blockWidth = width - LEFT_MARGIN + height;
+
+      const linesQuantity = Math.floor(((blockWidth) / (rotatedLineWidth + rotatedLineMargin)) * 2.1);
+
+      const lines = [...Array(linesQuantity)].map((i) => (
+        <View
+          key={i}
+          style={[
+            styles.blockedLine,
+            {
+              width: lineWidth,
+              marginRight: lineMargin,
+              height: rotatedHeight,
+              left: height * -1,
+            },
+          ]}
+        />
+      ));
+      return (
+        <View
+          style={[
+            styles.blockedContainer,
+            {
+              height,
+              top,
+              marginLeft: LEFT_MARGIN,
+            },
+          ]}
+          key={event.id}
+        >
+          {lines}
+        </View>
+      );
+    });
+  };
+
   return (
     <ScrollView
-      refreshControl={<RefreshControl refreshing={refreshing} onRefresh={onRefresh} />}
       contentContainerStyle={[styles.contentStyle, { width }]}
+      showsVerticalScrollIndicator={false}
+      contentOffset={{ y: contentOffset() }}
+      ref={scrollViewRef}
     >
+      {renderBlocks()}
       {renderLines()}
       {renderEvents()}
       {renderRedLine()}
